extracted DatePick class to separate date_pick_widget.dart file
